requirements.txt
ultralytics
cvzone
facenet-pytorch